checkpoint: renames "Attempt/Error/ActionableError.ts" to "Attempt/Error/Actionable.ts"
