+++ src/library/Attempt/Error/exports.ts
@@ -4,7 +4,7 @@ export {
 
 export {
   default as Attempt_Error_Actionable,
-} from './ActionableError';
+} from './Actionable';
 
 export type {
   default as Attempt_Error_Interpreter,
